% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Dataset}
\format{
An object of class \code{list} of length 6.
}
\usage{
simdata
}
\description{
Click \href{https://github.com/celehs/sureLDA/blob/master/data-raw/simdata.R}{HERE} to view details.
}
\examples{
str(simdata)
}
\keyword{datasets}
