% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sureLDA.R
\name{sureLDA}
\alias{sureLDA}
\title{Surrogate-guided ensemble Latent Dirichlet Allocation}
\usage{
sureLDA(X, weight, ICD, NLP, HU, filter, nEmpty = 20, alpha = 1,
  beta = 1, burnin = 50, ITER = 150)
}
\arguments{
\item{X}{nPatients x nFeatures matrix of feature counts}

\item{weight}{nPhenotypes x nFeatures matrix of phenotype-specific feature weights}

\item{ICD}{nPatients x nPhenotypes matrix of main ICD surrogate counts}

\item{NLP}{nPatients x nPhenotypes matrix of main NLP surrogate counts}

\item{HU}{nPatients-dimensional hospital utilization vector}

\item{filter}{nPatients x nPhenotypes binary matrix indicating filter-positives}

\item{nEmpty}{Number of 'empty' topics to include in LDA step}

\item{alpha}{LDA hyperparameters}

\item{beta}{LDA hyperparameters}

\item{burnin}{number of burnin Gibbs iterations}

\item{ITER}{number of subsequent iterations for inference}
}
\value{
scores = nPatients x nPhenotypes matrix of sureLDA phenotype scores; probs = nPatients x nPhenotypes matrix of posterior probabilities (post-clustering of sureLDA scores)
}
\description{
Surrogate-guided ensemble Latent Dirichlet Allocation
}
