% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sureLDA.R
\name{sureLDA}
\alias{sureLDA}
\title{Surrogate-guided ensemble Latent Dirichlet Allocation}
\usage{
sureLDA(
  X,
  ICD,
  NLP,
  HU,
  filter,
  prior = "PheNorm",
  weight = "beta",
  nEmpty = 20,
  alpha = 100,
  beta = 100,
  burnin = 50,
  ITER = 150,
  phi = NULL,
  labeled = NULL
)
}
\arguments{
\item{X}{nPatients x nFeatures matrix of feature counts}

\item{ICD}{nPatients x nPhenotypes matrix of main ICD surrogate counts}

\item{NLP}{nPatients x nPhenotypes matrix of main NLP surrogate counts}

\item{HU}{nPatients-dimensional hospital utilization vector}

\item{filter}{nPatients x nPhenotypes binary matrix indicating filter-positives}

\item{prior}{'PheNorm', 'MAP', or nPatients x nPhenotypes matrix of prior probabilities}

\item{weight}{'beta', 'uniform', or nPhenotypes x nFeatures matrix of feature weights}

\item{nEmpty}{Number of 'empty' topics to include in LDA step}

\item{alpha}{LDA hyperparameters}

\item{beta}{LDA hyperparameters}

\item{burnin}{number of burnin Gibbs iterations}

\item{ITER}{number of subsequent iterations for inference}

\item{phi}{...}

\item{labeled}{(optional) = NA when missing, non-NA for observed}
}
\description{
Surrogate-guided ensemble Latent Dirichlet Allocation
}
